GTD-1234 :bug: workspace duplication bug fixed
The problem was that when the previous postgres insert query run it threw an error when it tried to insert the data from the table of the source schema to an already created table in the destination schema.
The reason of the issue was that if we created a workspace long time ago, and we modified the table columns, like added a new column to the table and run a migration. The old table will be appended with that new column. This would be totally okay, but if the new column was not inserted to the end of the table definition in the tables.py, then the next time when the app tried to duplicate the workspace it created a new empty schema for the duplicated workspace and after that it created the tables by the table definitions defined in the tables.py.
Now this would result in a conflict because the order of the column in the old table (with migration) will be different from the new table (in the duplicated workspace schema). So the insert will fail because it will try to copy to different columns.