GTD-1258 :bug: migrations bug fixed
The problem was that the migration immediately throw an error as soon as P&G updated their Gitential application. The error was because of the migration with the revision_id=005. And the cause was the wrong postgresql query, and because it didn't check if the column exists before it tried to alter it.
The migration postgresql query might look like an overkill but noting else worked for me so far. PostgreSQL won't accept the IF EXISTS in the ALTER TABLE statement. Besides this DO block below, I had to set the isolation_level="AUTOCOMMIT" flag in the create_engine function call in the SQLGitentialBackend class in the /backends/sql/__init__.py otherwise the column name change is not working.