GTD-1268 :bug: SQLAlchemy bug fixed
isolation_level="AUTOCOMMIT" was removed from the initialization of the SQLAlchemy engine because it caused a lot of errors with stats_v2 queries in the database.
The reason to set this flag before because some sql queries in the migrations were not committed.
Solution: The isolation_level="AUTOCOMMIT" removed and the migration query execution was refactored.