GTD-1314 :sparkles: get organization repos for user implemented
This new feature is only implemented for onprem installations.
First scenario is when it is an onprem installation and no given user organization name came from the endpoint. In this way the software will get all of the organizations associated with a user so far and it will request the repositories for all of them one by one.
The second scenario is similar with the only difference that the user has to provide the organization names. If the user won't provide the names the first scenario will take into action.