GTD-1472 :bug: Fixed database error
An error come up for P&G when they tried to run the cleanup for all of their workspaces.
psycopg2.OperationalError: server closed the connection unexpectedly
Most probably it was the cause of the fact that our app did not wait for each workspace cleanup to finish, but tried to run the next process instantly.