GTD-1436 :bug: cleanup process corrected
From P&G's side there was an error in delete rows function in the extracted_commits table. The reason was because there were around 480K rows needed to delete. What I implemented was that delete rows will be applied in smaller chunks (max 100 rows) if the number of the rows to be deleted is more than 100.